--- a/Content.Shared/_NF/CCVar/NFCCVars.cs
+++ b/Content.Shared/_NF/CCVar/NFCCVars.cs
@@ -57,13 +57,13 @@ public sealed class NFCCVars
     /// The amount of time the bus waits at a station.
     /// </summary>
     public static readonly CVarDef<float> PublicTransitWaitTime =
-        CVarDef.Create("nf14.publictransit.wait_time", 180f, CVar.SERVERONLY);
+        CVarDef.Create("nf14.publictransit.wait_time", 120f, CVar.SERVERONLY);
 
     /// <summary>
     /// The amount of time the flies through FTL space.
     /// </summary>
     public static readonly CVarDef<float> PublicTransitFlyTime =
-        CVarDef.Create("nf14.publictransit.fly_time", 50f, CVar.SERVERONLY);
+        CVarDef.Create("nf14.publictransit.fly_time", 15f, CVar.SERVERONLY);
 
     /*
      *  World Gen