@Danchi299 2025-07-05
> transit time for 50 seconds to 15 seconds

But Public Transport Shuttle currently teleports instantly on live?